backend/requirements.txt
# FastAPI y servidor web
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0

# Base de datos
psycopg2-binary==2.9.9

# Autenticación
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6

# Procesamiento de imágenes y visión por computadora
opencv-python==4.8.1.78
ultralytics==8.0.196
easyocr==1.7.0
numpy==1.24.3
Pillow==10.1.0

# Utilidades
python-dotenv==1.0.0
pydantic==2.5.0
